refactor: Use standard `Dial`/`Listen` Go terminology for VSock
